+++ djpcms/media/djpcms/djpcms.js
@@ -446,6 +446,31 @@
 		}
 	});
 	
+	dj.addDecorator({
+		id:"color-picker",
+		decorate: function($this, config) {
+			$('input.color-picker', $this).each(function() {
+				var div = $('<div class="color-picker"></div>');
+				var iel = $(this).hide().after(div);
+				var v = iel.val();
+				div.append(iel.remove());
+				div.css('backgroundColor', '#' + v);
+				div.ColorPicker({
+					onSubmit: function(hsb, hex, rgb, el) {
+						var elem = $(el);
+						$('input',elem).val(hex);
+						elem.css('backgroundColor', '#' + hex);
+						elem.ColorPickerHide();
+					},
+					onBeforeShow: function () {
+						var v = $('input',this).val();
+						$(this).ColorPickerSetColor(v);
+					}
+				});
+			});
+		}
+	});
+	
 		
 })(jQuery);