+++ djpcms/plugins/text.py
@@ -7,7 +7,7 @@ from djpcms.conf import settings
 from djpcms.utils import json, form_kwargs, mark_safe
 from djpcms.models import SiteContent
 from djpcms.plugins import DJPplugin
-from djpcms.forms import SlugField
+from djpcms.forms import SlugField, LazyChoiceField
 from djpcms.utils.html import formlet
 from djpcms.markup import markuplib
 
@@ -35,9 +35,9 @@ class ChangeTextContent(forms.Form):
                                   label = 'New content unique code',
                                   help_text = 'When creating a new content give a unique name you like',
                                   required = False)
-    markup       = forms.ChoiceField(choices = markuplib.choices,
-                                     initial = markuplib.default,
-                                     required = False)
+    markup       = LazyChoiceField(choices = markuplib.choices,
+                                   initial = markuplib.default,
+                                   required = False)
     
     def __init__(self, *args, **kwargs):
         request = kwargs.pop('request',None)